handle ORGANIZED dir not exist

+++ autoganized/moveDirFromOrganized.py
@@ -4,9 +4,11 @@ import shutil
 
 def moveFrom_Organized(parentDir):
     rootDir = os.path.join(parentDir, 'ORGANIZED')
-    for item in os.listdir(rootDir):
-        shutil.move(os.path.join(rootDir, item), os.path.join(parentDir, item))
-    os.rmdir(rootDir)
+    
+    if os.path.isdir(rootDir):
+        for item in os.listdir(rootDir):
+            shutil.move(os.path.join(rootDir, item), os.path.join(parentDir, item))
+        os.rmdir(rootDir)
 
 def walkInDir(cwd):
     for item in os.scandir(cwd):
